Add workaround for JIT compiler to handle soft compiled methods
In-image Smalltalk compiler produces byte codes with trailing zeroes.
In case of soft VM this is not a problem, because control flow
never reaches them due to default method exit sequence.

In case of JIT compiler we need to fully process all available
bytecodes because we do not know how control will flow.

Correct solution will be to fix the in-image Smalltalk compiler.
However it is a complex task that should be handled carefully.

For now we're just ignoring such zero opcodes treating them as NOP.

Issue: #26