Adds BasicBlock reference API
This commit adds reference information to the BasicBlocks so that it knows
which basic blocks are referencing it by jumping to the first byte offset
of the current basic block. Information is collected during bytecode parsing.

Additionally, some functions are added for the TSmalltalkInstruction which
helps classifying instructions during iteration.

These functions are used in higher level control flow analysis API.

Issue: #32