Adds ControlGraph constructor which accepts ParsedMethod
ControlGraph is used not only for methods but also for inline blocks.
We need to generalize graph logic to allow analyzing both ParsedMethod
and ParsedBlock.

Use getParsedBytecode() method to get the associated bytecode.
Now getParsedMethod() return a container method for blocks and
the method itself if it is a method's graph.

Issue: #32