Phi nodes are now encoded in a separate pass
Encoding of phi nodes result in a complex code modification
in several basic blocks. If some of basic blocks will not be
finished at the time of phi encoding, malformed code will be produced.

That's why we encode phi nodes at the very end of method processing.
Dummy values are used to bind phi nodes to the instructions requesting
phi as an argument.

Issue: #32