Adds logic that answers a question whether produced value should be protected
We should protect the value by holder if consumer of this value is far away.
By far away we mean that it crosses the barrier of potential garbage collection.
For example if value is consumed right next to the point it was produced, then
protection is not neccessary. Otherwise if there are memory allocation points in
between then value protection is a must.

In order to find out whether value protection is required, we trace the possible
control- and data-flow paths and check what instructions are on the way between
value producer and all potential consumers. If there is at least one dangerous
operation on any of possible execution paths then the whole value should be protected.

Issue: #32