Adds special handling of messages that do not return normally
Some methods like Object>>error: do not return a value in a usual way.
In that case control flow is interrupted and all call chain aborts.

Type analyzer uses an empty composite type () to mark such special case.
Please note that () is not equal to * or ? types that still return a value.

Issue: #17
Issue: #92